Implement basic spot hold strategy
This commit adds a new contract, `SpotHold`, which implements a simple
strategy of buying and then selling an arbitrary ERC20 token via
Uniswap v3.

The intent is for this contract to serve as an implementation contract
for a given position and to be subject to the `Manager` contract's
control. This contract also serves as an example of what potential
implementations can look like.